Added schemes preview to own file
